Refactored supplement scripts
